+++ NEWS.md
@@ -4,9 +4,7 @@
 
 - New: Summarize the overall number of mutants generated, caught, missed, etc, at the end.
 
-- Fixed: Works properly with crates that have relative `path` dependencies in `Cargo.toml`, by rewriting them to absolute paths in the scratch directory. (`patch` and `replace` configurations, and `.cargo/config.toml` are not yet supported.)
-
-- New: Summarize the overall number of mutants generated, caught, missed, etc, at the end.
+- Fixed: Works properly with crates that have relative `path` dependencies in `Cargo.toml` or `.cargo/config.toml`, by rewriting them to absolute paths in the scratch directory.
 
 ## 0.2.7